requirements.txt
fastapi
uvicorn[standard]
pydantic
python-dotenv
pandas
requests
sqlalchemy
psycopg2-binary
xlsxwriter
